+++ acl/src/main/java/org/apache/rocketmq/acl/plug/AclPlugRuntimeException.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.acl.plug.exception;
+package org.apache.rocketmq.acl.plug;
 
 public class AclPlugRuntimeException extends RuntimeException {
