+++ acl/src/main/java/org/apache/rocketmq/acl/plug/AuthenticationResult.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.acl.plug.entity;
+package org.apache.rocketmq.acl.plug;
 
 public class AuthenticationResult {
