+++ acl/src/main/java/org/apache/rocketmq/acl/plug/BorkerAccessControl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.acl.plug.entity;
+package org.apache.rocketmq.acl.plug;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -556,8 +556,8 @@ public class BorkerAccessControl extends AccessControl {
     public String toString() {
         StringBuilder builder = new StringBuilder();
         builder.append("BorkerAccessControl [permitSendTopic=").append(permitSendTopic).append(", noPermitSendTopic=")
-                .append(noPermitSendTopic).append(", permitPullTopic=").append(permitPullTopic)
-                .append(", noPermitPullTopic=").append(noPermitPullTopic);
+            .append(noPermitSendTopic).append(", permitPullTopic=").append(permitPullTopic)
+            .append(", noPermitPullTopic=").append(noPermitPullTopic);
         if (!!sendMessage)
             builder.append(", sendMessage=").append(sendMessage);
         if (!!sendMessageV2)
