+++ acl/src/main/java/org/apache/rocketmq/acl/plug/NetaddressStrategy.java
@@ -14,9 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.acl.plug.strategy;
-
-import org.apache.rocketmq.acl.plug.entity.AccessControl;
+package org.apache.rocketmq.acl.plug;
 
 public interface NetaddressStrategy {
