+++ acl/src/test/java/org/apache/rocketmq/acl/plug/NetaddressStrategyTest.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.acl.plug.strategy;
+package org.apache.rocketmq.acl.plug;
 
-import org.apache.rocketmq.acl.plug.entity.AccessControl;
-import org.apache.rocketmq.acl.plug.exception.AclPlugRuntimeException;
 import org.junit.Assert;
 import org.junit.Test;
